migrate request headerList
