+++ src/app/views/common/lazy-loader/component-registry/index.tsx
@@ -8,7 +8,7 @@ import LazyGraphToolkit from  '../../../query-response/graph-toolkit/GraphToolki
 import LazySnippets from  '../../../query-response/snippets/Snippets';
 import LazyCopyButton from  '../../copy-button/CopyButtonV9';
 import LazyAuth from  '../../../query-runner/request/auth/AuthV9';
-import LazyRequestHeaders from  '../../../query-runner/request/headers/RequestHeadersV9';
+import LazyRequestHeaders from  '../../../query-runner/request/headers/RequestHeaders';
 import LazyHistory from  '../../../sidebar/history/History';
 import LazyResourceExplorer from  '../../../sidebar/resource-explorer/ResourceExplorer';
