+++ src/app/views/query-runner/request/headers/Headers.styles.ts
@@ -1,25 +1,25 @@
-import { FontSizes } from '@fluentui/react';
+import { makeStyles } from '@fluentui/react-components';
 
-export const headerStyles = () => {
-  return {
-    container: {
-      textAlign: 'center',
-      padding: 10,
-      overflowY: 'auto',
-      overflowX: 'hidden'
-    },
-    itemContent: {
-      marginTop: '2.5%'
-    },
-    rowContainer: {
-      fontSize: FontSizes.medium,
-      position: 'relative'
-    },
-    detailsRow: {
-      display: 'flex',
-      justifyContent: 'center',
-      alignItems: 'center'
-    },
-    headersList: { marginBottom: '120px' }
-  };
-};
+export const useHeaderStyles = makeStyles({
+  container: {
+    overflow: 'auto hidden',
+    padding: '10px',
+    textAlign: 'start',
+    height: '100%'
+  },
+  itemContent: {
+    marginTop: '2.5%'
+  },
+  rowContainer: {
+    fontSize: '1rem',
+    position: 'relative'
+  },
+  detailsRow: {
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center'
+  },
+  actionButtons: {
+    display: 'flex'
+  }
+});
